[refactor] 실패하던 테스트 수 (#399)
